pyproject.toml
[tool.poetry]
name = "curation-buddy"
version = "0.1.0"
description = "Don't bug your friends with articles they'll never read. Use the curation buddy to engage socially with your ideas. AI's have infinite attention, leverage them instead!"
authors = ["vintro <vince@plasticlabs.ai>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
py-cord = "^2.5.0"
honcho-ai = "^0.0.4"
langchain-core = "^0.1.29"
openai = "^1.13.3"
langchain-openai = "^0.0.8"
python-dotenv = "^1.0.1"
realtime = "^1.0.2"
langchain-community = "^0.0.26"
apify-client = "^1.6.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
